// frontend/src/lib/api/api.interfaces.d.ts
export interface IGetUserResponse {
    email: string;
    full_name: string;
    disabled: boolean;
    auth_method: 'email' | 'google';
    photo: string;
    id: string;
    has_api_key: boolean;
}
// ...
export interface Organization {
    id: string;
    name: string;
    logo: string | null;
    role: OrganizationRoleEnum;
    created_at: Date;
    updated_at: Date;
    members_count: number;
    permissions: OrganizationPermissions;
}
export interface Invitation {
    id: string;
    email: string;
    role: OrganizationRoleEnum;
    organization_id: string;
    invited_by_user: Omit<IGetUserResponse, 'auth_method'>;
    created_at: Date;
    organization: Omit<Organization, 'role'>;
}
export interface CreateOrganizationPayload {
    name: string;
    logo: string;
}
export interface CreateOrganizationResponse {
    id: string;
    name: string;
    logo: string;
    created_at: Date;
    updated_at: Date;
}
export interface OrganizationDetails extends Omit<Organization, 'role'> {
    invites: Omit<Invitation, 'organization'>[];
    members: Member[];
    orbits: Orbit[];
    members_limit: number;
    orbits_limit: number;
    total_orbits: number;
    total_members: number;
    members_by_role: Record<OrganizationRoleEnum, number>;
    satellites_limit: number;
    artifacts_limit: number;
    total_satellites: number;
    total_artifacts: number;
}
export interface Member {
    id: string;
    organization_id: string;
    role: OrganizationRoleEnum;
    user: Omit<IGetUserResponse, 'auth_method'>;
    created_at: Date;
    updated_at: Date | null;
}
export interface AddMemberPayload {
    user_id: string;
    organization_id: string;
    role: OrganizationRoleEnum;
}
export interface UpdateMemberPayload {
    role: OrganizationRoleEnum;
}
export interface CreateInvitePayload {
    email: string;
    role: OrganizationRoleEnum;
    organization_id: string;
}
export interface Orbit {
    id: string;
    name: string;
    organization_id: string;
    total_members: number;
    created_at: Date;
    updated_at: Date | null;
    bucket_secret_id: string;
    total_collections: number;
    role: OrbitRoleEnum;
    permissions: OrbitPermissions;
    total_artifacts: number;
}
// ...
export interface OrganizationPermissions {
    organization: [
        PermissionEnum.read,
        PermissionEnum.update,
        PermissionEnum.delete,
        PermissionEnum.leave
    ];
    organization_user: Omit<PermissionEnum, PermissionEnum.deploy>;
    organization_invite: Omit<PermissionEnum, PermissionEnum.update & PermissionEnum.deploy>;
    billing: [PermissionEnum.read, PermissionEnum.update];
    orbit: [PermissionEnum.create];
}
export interface OrbitPermissions {
    orbit: Omit<PermissionEnum, PermissionEnum.deploy>;
    orbit_user: Omit<PermissionEnum, PermissionEnum.deploy>;
    artifact: PermissionEnum;
    collection: PermissionEnum;
    track: Omit<PermissionEnum, PermissionEnum.deploy>;
}
export declare enum PermissionEnum {
    list = "list",
    read = "read",
    create = "create",
    update = "update",
    delete = "delete",
    deploy = "deploy",
    leave = "leave"
}